section 10
seccion 10 de 193 a 196
